Move files and fix tests with date
