+++ packages/plugins/i18n/admin/src/components/CMEditViewInjectedComponents/CMEditViewCopyLocale/utils/tests/removePasswordAndRelationsFieldFromData.test.js
