+++ packages/plugins/i18n/admin/src/components/CMEditViewInjectedComponents/CMEditViewLocalePicker/Option.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import get from 'lodash/get';
 import { Flex, Text } from '@buffetjs/core';
 import { RelationDPState } from '@strapi/helper-plugin';
-import { getTrad } from '../../utils';
+import { getTrad } from '../../../utils';
 
 const TextGrow = styled(Text)`
   flex-grow: 2;
